+++ cli/cli/usage_nopflag.go
@@ -1,0 +1,9 @@
+// +build !pflag
+
+package usage
+
+import flag "github.com/spf13/pflag"
+
+func (c *CLI) additionalFlagSets() map[string]*flag.FlagSet {
+	return nil
+}