+++ daemon/daemon.go
@@ -5,7 +5,7 @@ package daemon
 import (
 	"os"
 
-	"github.com/akutz/gofig"
+	gofig "github.com/akutz/gofig/types"
 	apitypes "github.com/codedellemc/libstorage/api/types"
 
 	"github.com/codedellemc/rexray/daemon/module"